auth_ephemeral: Rename module interface file #82
